all routes protected by their respective permissions
- for example
routes like 'students', 'students/1' will be protected by view-students permission and so on....
- from now on to add a new crud you just need to add the plural version of the route(eg departments) mapped with the singular but capitalized(Department) in the $resourceRoutes array.

+++ routes/web.php
@@ -1,19 +1,8 @@
 <?php
 
-use App\Http\Controllers\DepartmentController;
-use App\Http\Controllers\EmployeeController;
-use App\Http\Controllers\InventoryCategoryController;
-use App\Http\Controllers\InventoryController;
-use App\Http\Controllers\InventorySupplierController;
 use App\Http\Controllers\ProfileController;
-use App\Http\Controllers\StudentController;
-use App\Http\Controllers\RoleController;
-use Illuminate\Foundation\Application;
 use App\Http\Controllers\PermissionController;
-use App\Http\Controllers\ProgramController;
-use App\Http\Controllers\TenantController;
-use App\Http\Controllers\AcademicDocumentController;
-use App\Http\Controllers\InstructorController;
+use App\Http\Controllers\RoleController;
 use Illuminate\Support\Facades\Route;
 use Inertia\Inertia;
 
@@ -22,33 +11,64 @@ Route::middleware('auth')->group(function () {});
 
 Route::middleware(['auth'])->group(function () {
     // Dashboard
-    Route::get('/', function () {
-        return Inertia::render('Dashboard');
-    })->name('dashboard');
+    Route::get('/', fn() => Inertia::render('Dashboard'))->name('dashboard');
 
     // Profiles related routes
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
-    // Role and Permission Related routes
-    Route::resource('roles', RoleController::class);
-    Route::resource('permissions', PermissionController::class);
-    Route::get('/roles/{role}/permissions', [RoleController::class, 'assign'])->name('roles.permissions');
-    Route::put('/roles/{role}/permissions', [RoleController::class, 'attach'])->name('roles.attach');
-    Route::delete('/roles/{role}/permissions/{permission}', [RoleController::class, 'detach'])->name('roles.detach');
+    // Role and Permission Routes
+    Route::middleware(['can:view-roles'])->resource('roles', RoleController::class);
+    Route::middleware(['can:view-permissions'])->resource('permissions', PermissionController::class);
 
-    // Main Application crud resource routes
-    Route::resource('students', StudentController::class);
-    Route::resource('departments', DepartmentController::class);
-    Route::resource('programs', ProgramController::class);
-    Route::resource('employees', EmployeeController::class);
-    Route::resource('inventories', InventoryController::class);
-    Route::resource('inventorySuppliers', InventorySupplierController::class);
-    Route::resource('inventoryCategories', InventoryCategoryController::class);
-    Route::resource('instructors', InstructorController::class);
-    Route::resource('academic-documents', AcademicDocumentController::class);
-    Route::resource('tenants', TenantController::class);
+    Route::get('/roles/{role}/permissions', [RoleController::class, 'assign'])->middleware('can:assign-permissions-roles')->name('roles.permissions');
+    Route::put('/roles/{role}/permissions', [RoleController::class, 'attach'])->middleware('can:attach-permissions-roles')->name('roles.attach');
+    Route::delete('/roles/{role}/permissions/{permission}', [RoleController::class, 'detach'])->middleware('can:detach-permissions-roles')->name('roles.detach');
+
+    // CRUD Resources with Permissions
+    // $controllers = [
+    //     'students' => StudentController::class,
+    //     'departments' => DepartmentController::class,
+    //     'programs' => ProgramController::class,
+    //     'employees' => EmployeeController::class,
+    //     'inventories' => InventoryController::class,
+    //     'inventorySuppliers' => InventorySupplierController::class,
+    //     'inventoryCategories' => InventoryCategoryController::class,
+    //     'instructors' => InstructorController::class,
+    //     'academic-documents' => AcademicDocumentController::class,
+    //     'tenants' => TenantController::class,
+    // ];
+
+    $resourceRoutes = [
+        'departments' => 'Department',
+        'students' => 'Student',
+        'programs' => 'Program',
+        'employees' => 'Employee',
+        'inventories' => 'Inventory',
+        'inventorySuppliers' => 'InventorySupplier',
+        'inventoryCategories' => 'InventoryCategory',
+        'instructors' => 'Instructor',
+        'academic-documents' => 'AcademicDocument',
+        'tenants' => 'Tenant',
+    ];
+
+    foreach ($resourceRoutes as $route => $singular) {
+        $controller = "App\\Http\\Controllers\\{$singular}Controller";
+        $singularToLower = strtolower($singular);
+
+        Route::middleware("can:create-$route")->post("$route", [$controller, 'store'])->name("$route.store");
+        Route::middleware("can:create-$route")->get("$route/create", [$controller, 'create'])->name("$route.create");
+        
+        Route::middleware("can:view-$route")->get("$route", [$controller, 'index'])->name("$route.index");
+        Route::middleware("can:view-$route")->get("$route/{{$singularToLower}}", [$controller, 'show'])->name("$route.show");
+        
+        Route::middleware("can:update-$route")->put("$route/{{$singularToLower}}", [$controller, 'update'])->name("$route.update");
+        Route::middleware("can:update-$route")->get("$route/{{$singularToLower}}/edit", [$controller, 'edit'])->name("$route.edit");
+
+        Route::middleware("can:delete-$route")->delete("$route/{{$singularToLower}}", [$controller, 'destroy'])->name("$route.destroy");
+    }
 });
+
 
 require __DIR__ . '/auth.php';
